// frontend/src/features/factory/fg-receipt-summary/types.ts
export interface IFgReceiptSummary {
  created_at: string;
  grade: string;
  product_type_code: string;
  product_type_id: number;
  product_type_name: string;
  production_date: string;
  production_order_code: string;
  production_order_id: number;
  production_order_name: string;
  rubber_block_code: string;
  rubber_block_id: number;
  status: string;
  updated_at: string;
  weight: string;
}